requirements.txt
python-telegram-bot==20.8
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0 